scenarios/handson_iot_hub/pyproject.toml
[tool.poetry]
name = "handson-iot-hub"
version = "0.0.1"
description = "handson IoT Hub"
authors = ["ks6088ts <ks6088ts@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.9.0"
python-dotenv = "^1.0.1"
tenacity = "^8.2.3"
azure-iot-hub = "^2.6.1"
azure-iot-device = "^2.13.0"
azure-storage-blob = "^12.19.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
